graphics and card deal added
Graphics are added in svg form. The cards are now randomly dealt,
correctly, I believe. There are three classes: card and deck and
mainWindow. The cards are contained in deck, as you might guess.  The
window deals with the graphics. If window needs access to the
information contained in deck or card, it first passes the deck
pointer, and then may access the cards through the deck method
“getCards”.