all buttons are created now
All buttons are created, including both players’ hands, their captured
cards, the field, and the deck. The captured cards are invisible at the
moment. Some of the field cards are invisible. None of the buttons
work, except for the players’ cards, which only play the cards clicked,
and having them merely go to the table.